pyproject.toml
[tool.poetry]
name = "python-project-lvl3"
version = "0.1.0"
description = ""
authors = ["Abdu <abdujabbar.mirkhalikov@gmail.com>"]

packages = [
  { include = "page_loader" },
]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
flake8 = "^4.0.1"
bs4 = "^0.0.1"
pytest = "^7.1.2"
validators = "^0.19.0"
vcrpy = "^4.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
page-loader = "page_loader.scripts.page_loader:main"



